Fix secure agg
