+++ federatedml/transfer_variable/auth_conf/table_aggregator_trans_var.json
@@ -1,5 +1,5 @@
 {
-  "federatedml.framework.homo.procedure.table_aggregator.TableScatterTransVar": {
+  "federatedml.framework.hetero.procedure.table_aggregator.TableScatterTransVar": {
     "client_table": {
       "src": [
         "guest",
@@ -10,6 +10,78 @@
       ]
     },
     "server_table": {
+      "src": [
+        "arbiter"
+      ],
+      "dst": [
+        "guest",
+        "host"
+      ]
+    },
+    "ModelScatterTransVar.client_model": {
+      "src": [
+        "guest",
+        "host"
+      ],
+      "dst": [
+        "arbiter"
+      ]
+    },
+    "ModelBroadcasterTransVar.server_model": {
+      "src": [
+        "arbiter"
+      ],
+      "dst": [
+        "guest",
+        "host"
+      ]
+    },
+        "AggregatorTransVar.ModelScatterTransVar.client_model": {
+      "src": [
+        "guest",
+        "host"
+      ],
+      "dst": [
+        "arbiter"
+      ]
+    },
+    "AggregatorTransVar.ModelBroadcasterTransVar.server_model": {
+      "src": [
+        "arbiter"
+      ],
+      "dst": [
+        "guest",
+        "host"
+      ]
+    },
+    "RandomPaddingCipherTransVar.UUIDTransVar.uuid": {
+      "src": [
+        "arbiter"
+      ],
+      "dst": [
+        "guest",
+        "host"
+      ]
+    },
+    "RandomPaddingCipherTransVar.DHTransVar.p_power_r": {
+      "src": [
+        "guest",
+        "host"
+      ],
+      "dst": [
+        "arbiter"
+      ]
+    },
+    "RandomPaddingCipherTransVar.DHTransVar.p_power_r_bc": {
+      "src": [
+        "arbiter"
+      ],
+      "dst": [
+        "guest",
+        "host"
+      ]
+    },
+    "RandomPaddingCipherTransVar.DHTransVar.pubkey": {
       "src": [
         "arbiter"
       ],
